requirements/requirements.txt
awesome-slugify==1.6.5
blinker==1.4
Flask==0.10.1
Flask-BabelEx==0.9.2
Flask-Cache==0.13.1
Flask-Gravatar==0.4.2
Flask-Mistune==0.1.1
Flask-OAuthlib==0.9.2
# Flask-Login broke compatibility
flask-security==1.7.4  # rq.filter: <1.7.5
# Pillow 3.0.0 raises --enable-zlib and --enable-jpeg error on some O.S
Pillow==2.9.0  # rq.filter:<3.0.0
PyRSS2Gen==1.1
requests==2.9.0
pyshorteners==0.5.8
cached-property==1.3.0
shiftpy==0.1.3
click==6.2

# tempfix
mongoengine==0.10.1  # rq.filter:<0.10.2

# mongoengine is not working yet with pymongo3
pymongo==2.9.1  # rq.filter: <3

# There is a compatibility break in Flask-Login 0.3
Flask-Login==0.2.11 # rq.filter: <0.3

# Mistune 0.7.1 broke base64; inline images
mistune==0.6

flask-admin
flask-mongoengine
